Refactor duplicated methods and constants in duck/credentials
