+++ pkg/apis/duck/validations.go
@@ -21,12 +21,21 @@ import (
 	"context"
 	"fmt"
 	"math"
+	"regexp"
 	"strconv"
 
 	"github.com/google/go-cmp/cmp"
 
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/equality"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"knative.dev/pkg/apis"
+)
+
+var (
+	// The name of a k8s ServiceAccount object must be a valid DNS subdomain name.
+	// https://kubernetes.io/docs/concepts/overview/working-with-objects/names/#dns-subdomain-names
+	ksaValidationRegex = regexp.MustCompile(`^[A-Za-z0-9](?:[A-Za-z0-9\-]{0,61}[A-Za-z0-9])?$`)
 )
 
 // ValidateAutoscalingAnnotations validates the autoscaling annotations.
@@ -93,4 +102,42 @@ func CheckImmutableClusterNameAnnotation(current *metav1.ObjectMeta, original *m
 		}
 	}
 	return errs
+}
+
+// ValidateCredential checks secret and service account.
+func ValidateCredential(secret *corev1.SecretKeySelector, kServiceAccountName string) *apis.FieldError {
+	if secret != nil && !equality.Semantic.DeepEqual(secret, &corev1.SecretKeySelector{}) && kServiceAccountName != "" {
+		return &apis.FieldError{
+			Message: "Can't have spec.serviceAccountName and spec.secret at the same time",
+			Paths:   []string{""},
+		}
+	} else if secret != nil && !equality.Semantic.DeepEqual(secret, &corev1.SecretKeySelector{}) {
+		return validateSecret(secret)
+	} else if kServiceAccountName != "" {
+		return validateK8sServiceAccount(kServiceAccountName)
+	}
+	return nil
+}
+
+func validateSecret(secret *corev1.SecretKeySelector) *apis.FieldError {
+	var errs *apis.FieldError
+	if secret.Name == "" {
+		errs = errs.Also(apis.ErrMissingField("secret.name"))
+	}
+	if secret.Key == "" {
+		errs = errs.Also(apis.ErrMissingField("secret.key"))
+	}
+	return errs
+}
+
+func validateK8sServiceAccount(kServiceAccountName string) *apis.FieldError {
+	match := ksaValidationRegex.FindStringSubmatch(kServiceAccountName)
+	if len(match) == 0 {
+		return &apis.FieldError{
+			Message: fmt.Sprintf(`invalid value: %s, serviceAccountName should have format: ^[A-Za-z0-9](?:[A-Za-z0-9\-]{0,61}[A-Za-z0-9])?$`,
+				kServiceAccountName),
+			Paths: []string{"serviceAccountName"},
+		}
+	}
+	return nil
 }
