+++ pkg/apis/messaging/v1alpha1/channel_validation.go
@@ -24,8 +24,6 @@ import (
 	"github.com/google/go-cmp/cmp"
 
 	"knative.dev/pkg/apis"
-
-	duckv1alpha1 "github.com/google/knative-gcp/pkg/apis/duck/v1alpha1"
 )
 
 func (c *Channel) Validate(ctx context.Context) *apis.FieldError {
@@ -45,7 +43,7 @@ func (cs *ChannelSpec) Validate(ctx context.Context) *apis.FieldError {
 		}
 	}
 
-	if err := duckv1alpha1.ValidateCredential(cs.Secret, cs.ServiceAccountName); err != nil {
+	if err := duck.ValidateCredential(cs.Secret, cs.ServiceAccountName); err != nil {
 		errs = errs.Also(err)
 	}
